ft_transcendence/django/requirements.txt
pycparser==2.21
cffi==1.16.0
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
django_rest_framework==0.1.0
Pillow==10.1.0
channels[daphne]==4.0.0
channels_redis==4.1.0
django-sslserver==0.22
daphne
Twisted[tls,http2]
pyopenssl
bcrypt
web3